Move setWithoutBalance method to abstract class -remove setWithoutBalance in trees implementations -add setWithoutBalance common implementation to BSTreeTemplate class -add abstract fabric method in BSTreeTemplate abstract class
