+++ lib/src/main/kotlin/trees/implementations/SimpleTree.kt
@@ -3,6 +3,11 @@ package trees.implementations
 import trees.templates.BSTreeTemplate
 
 class SimpleTree<K : Comparable<K>, V> : BSTreeTemplate<K, V, SimpleVertex<K, V>>() {
+
+    override fun fabricVertex(key: K, value: V): SimpleVertex<K, V> {
+        return SimpleVertex(key, value)
+    }
+
     override fun remove(key: K): V? {
         val vertex = vertByKey(key) ?: return null
         val parent = vertex.parent
@@ -41,39 +46,11 @@ class SimpleTree<K : Comparable<K>, V> : BSTreeTemplate<K, V, SimpleVertex<K, V>
     }
 
     override operator fun set(key: K, value: V): V? {
-        if (root == null) {
-            val vertex = SimpleVertex(key, value)
-            root = vertex
+        val (currentVert, oldValue) = setWithoutBalance(key, value, ::fabricVertex)
+        if (oldValue == null) {
             size += 1
-            return null
         }
-
-        var current = root ?: throw IllegalStateException("Can't be null, there was a check above")
-        while (true) {
-            val result = key.compareTo(current.key)
-            if (result < 0) {
-                if (current.left == null) {
-                    val newVertex = SimpleVertex(key, value)
-                    current.left = newVertex
-                    newVertex.parent = current
-                    size += 1
-                    return null
-                }
-                current = current.left ?: throw IllegalStateException("Case when cur.left is null is processed above")
-            } else if (result > 0) {
-                if (current.right == null) {
-                    val newVertex = SimpleVertex(key, value)
-                    current.right = newVertex
-                    newVertex.parent = current
-                    size += 1
-                    return null
-                }
-                current = current.right ?: throw IllegalStateException("Case when cur.right is null is processed above")
-            } else {
-                val oldValue = current.value
-                current.value = value
-                return oldValue
-            }
-        }
+        return oldValue
     }
+
 }
