+++ lib/src/main/kotlin/trees/implementations/Vertices.kt
@@ -15,7 +15,7 @@ class AVLVertex<K : Comparable<K>, V>(key: K, value: V) : VertexTemplate<K, V, A
 }
 
 class RBVertex<K : Comparable<K>, V>(key: K, value: V) : VertexTemplate<K, V, RBVertex<K, V>>(key, value) {
-    var color: Color = Color.BLACK
+    var color: Color = Color.RED
     var additionalType = false
 
     enum class Color {
